Moved the DSL code to TextualSummaryHelper
