+++ app/helpers/physical_storage_helper/textual_summary.rb
@@ -1,5 +1,4 @@
 module PhysicalStorageHelper::TextualSummary
-  include TextualMixins::NewTextualSummary
 
   def textual_group_list
     textual_summary do
